Update project structure to make the project more moudlar
