+++ src/resolvers/index.ts
@@ -1,0 +1,8 @@
+// Operations
+export { Query } from './Query';
+export { Mutation } from './Mutation';
+
+// Types
+export { Comment } from './Comment';
+export { Post } from './Post';
+export { User } from './User';
